Add support to always take a snapshot on run
Previously, running via --run with a scheduled snapshot it
is wrongly missing every other snapshot as it is usually a
second too young to be considered stale.

This change adds a --forcerun option to always take a
snapshot (irrelevant of age) and then purge any snapshots due
to be purged as usual.

fixes skymill/automated-ebs-snapshots#23

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>